feat: 공통 table 컴포넌트 내부 badge 공통 컴포넌트로 분리
