+++ src/components/ui/badge/Badge.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import { StyledBadge, StyledBadgeTextSpan } from './Badge.styles';
+import type { BadgeProps } from './types';
+
+export const Badge: React.FC<BadgeProps> = ({ $badgeColor, children }) => {
+  return (
+    <StyledBadge $badgeColor={$badgeColor}>
+      <StyledBadgeTextSpan>{children}</StyledBadgeTextSpan>
+    </StyledBadge>
+  );
+};
